fix rotation

+++ bnp/scene/visualize.py
@@ -16,4 +16,7 @@ def put_cubes(positions, prefix="debug", size=0.015, sampling_rate=1):
         bpy.ops.mesh.primitive_cube_add(size=size, location=(v[0], v[1], v[2]))
         bpy.context.object.name = f"debug_{str(idx)}"
         debug_collection.objects.link(bpy.context.object)
-        bpy.data.collections["Collection"].objects.unlink(bpy.context.object)
+        if bpy.context.object in list(bpy.data.collections["Collection"].objects):
+            bpy.data.collections["Collection"].objects.unlink(bpy.context.object)
+        if bpy.context.object in list(bpy.context.scene.collection.objects):
+            bpy.context.scene.collection.objects.unlink(bpy.context.object)
